+++ mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/adapter/PaginationableAdapter.kt
@@ -24,15 +24,15 @@ import android.widget.TextView
 import androidx.core.view.updateLayoutParams
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import ru.surfstudio.android.core.mvi.sample.R
-import ru.surfstudio.android.easyadapter.pagination.BasePaginationableAdapter
-import ru.surfstudio.android.easyadapter.pagination.PaginationState
+import ru.surfstudio.android.pagination.PaginationState
+import ru.surfstudio.android.pagination.EasyPaginationAdapter
 
 /**
  * Origin <a href="http://google.com">https://github.com/MaksTuev/EasyAdapter/tree/master/sample/src/main/java/ru/surfstudio/easyadapter/sample</a>
  *
  * Класс адаптера с поддержкой пагинации на основе EasyAdapter
  */
-open class PaginationableAdapter(onShowMoreListener: () -> Unit) : BasePaginationableAdapter() {
+open class PaginationableAdapter(onShowMoreListener: () -> Unit) : EasyPaginationAdapter() {
     protected var paginationFooterItemController: BasePaginationFooterController<*>? = null
 
     init {