+++ mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/screen/reactor_based/list/ComplexListEvent.kt
@@ -6,7 +6,7 @@ import ru.surfstudio.android.core.mvi.sample.ui.base.middleware.experimental.pag
 import ru.surfstudio.android.core.mvp.binding.rx.request.Request
 import ru.surfstudio.android.core.ui.state.LifecycleStage
 import ru.surfstudio.android.datalistpagecount.domain.datalist.DataList
-import ru.surfstudio.android.easyadapter.pagination.PaginationState
+import ru.surfstudio.android.pagination.PaginationState
 
 /**
  * События экрана [ComplexListActivityView]