+++ mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/screen/reducer_based/kitties/util/RequestMappers.kt
@@ -7,7 +7,7 @@ import ru.surfstudio.android.core.mvi.ui.mapper.RequestLoadingMapper
 import ru.surfstudio.android.core.mvp.binding.rx.request.data.*
 import ru.surfstudio.android.core.mvp.error.ErrorHandler
 import ru.surfstudio.android.datalistlimitoffset.domain.datalist.DataList
-import ru.surfstudio.android.easyadapter.pagination.PaginationState
+import ru.surfstudio.android.pagination.PaginationState
 import ru.surfstudio.android.rx.extension.scheduler.MainThreadImmediateScheduler
 
 /**