+++ template/RELEASE_NOTES.md
@@ -39,7 +39,7 @@
 * ANDDEP-982 other libraries versions changed;
 * ANDDEP-915 Added `validateCrossFeatureRoutes.gradle.kts` task. **Note**: supported only `.kt` source files parsing;
 * ANDDEP-599 Added `CompletableSafeConverter` to parse not empty Completable-request response;
-
+* ANDDEP-1038 Fix import for EasyPaginationAdapter
 ## 0.4.0
 ##### Template
 * Updated navigation mechanisms in template